agent/requirements.txt
# ============================================================
# SAP Spektra Agent — Dependencias
# ============================================================
# Instalar con: pip install -r requirements.txt

# ── Siempre requeridos ──

# Métricas del sistema operativo (CPU, RAM, disco, red)
psutil>=5.9.0

# Peticiones HTTP (enviar datos al API, sapcontrol SOAP, cloud metadata)
requests>=2.31.0

# Leer archivos de configuración YAML
pyyaml>=6.0

# Programar tareas periódicas (recolección cada N segundos)
schedule>=1.2.0

# Parsear respuestas XML de sapcontrol
lxml>=4.9.0

# Encriptación de credenciales locales (AES Fernet + PBKDF2)
cryptography>=41.0.0


# ── Opcionales: SDKs Cloud (para operaciones HA/DR) ──
# Descomentar SOLO el que corresponda al proveedor cloud:

# AWS (EC2 resize, EIP, Route53, CloudFormation, S3)
# boto3>=1.34.0

# Azure (VM resize, Traffic Manager, Managed Disks)
# azure-identity>=1.15.0
# azure-mgmt-compute>=30.0.0
# azure-mgmt-network>=25.0.0

# GCP (Instance resize, Cloud DNS)
# google-auth>=2.25.0
# google-cloud-compute>=1.15.0
# google-cloud-dns>=0.35.0


# ── Opcionales: Drivers de base de datos ──
# Descomentar SOLO el que corresponda al sistema:

# SAP HANA
# hdbcli>=2.17.0

# Oracle
# oracledb>=2.0.0

# IBM DB2
# ibm_db>=3.2.0

# Microsoft SQL Server / SAP ASE (mismo driver)
# pymssql>=2.2.0

# SAP MaxDB → no necesita driver, usa dbmcli (viene con MaxDB)
